Prueba con jpa

+++ src/main/java/com/alexander/proyecto/services/PersonaService.java
@@ -3,6 +3,7 @@ package com.alexander.proyecto.services;
 import com.alexander.proyecto.dto.PersonaRequest;
 import com.alexander.proyecto.dto.PersonaResponse;
 import com.alexander.proyecto.entities.Persona;
+import com.alexander.proyecto.enums.Genero;
 
 import java.util.List;
 
@@ -13,7 +14,7 @@ public interface PersonaService {
 
     List<PersonaResponse> obtenerPorEmail(String email);
 
-    List<PersonaResponse> obtenerPorRangoEdad(int edadInicio, int edadFin);
+    List<PersonaResponse> obtenerPorRangoEdad(Short edadInicio, Short edadFin);
 
     List<PersonaResponse> obtenerPorTelefono(String telefono);
